requirements.txt
boto3==1.34.113
clickhouse_connect==0.7.11
prometheus_client==0.20.0
PyYAML==5.4.1
setuptools==59.6.0
SQLAlchemy==2.0.30
